Renamed settings to storage
